refactor(factoryinsight): merged controllers for endpoints deeper than `/:dataFormat`
