+++ golang/cmd/factoryinsight/v2/models/dto-productionline.go
@@ -8,3 +8,9 @@ type ProductionLine struct {
 type GetProductionLinesResponse struct {
 	ProductionLines []ProductionLine `json:"productionLines"`
 }
+
+type GetProductionLineRequest struct {
+	EnterpriseName string `uri:"enterpriseName" binding:"required"`
+	SiteName       string `uri:"siteName" binding:"required"`
+	AreaName       string `uri:"areaName" binding:"required"`
+}
