--- golang/cmd/factoryinsight/v2/models/dto-taggroup.go
@@ -1,10 +1,0 @@
-package models
-
-type TagGroup struct {
-	Id   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-type GetTagGroupsResponse struct {
-	TagGroups []TagGroup `json:"tagGroups"`
-}
