+++ golang/cmd/factoryinsight/v2/models/dto-workcell.go
@@ -8,3 +8,10 @@ type WorkCell struct {
 type GetWorkCellsResponse struct {
 	WorkCells []WorkCell `json:"workCells"`
 }
+
+type GetWorkCellsRequest struct {
+	EnterpriseName     string `uri:"enterpriseName" binding:"required"`
+	SiteName           string `uri:"siteName" binding:"required"`
+	AreaName           string `uri:"areaName" binding:"required"`
+	ProductionLineName string `uri:"productionLineName" binding:"required"`
+}
